+++ src/features/reply/components/rereply-item.tsx
@@ -1,6 +1,6 @@
-import { ReplyContent } from '@/components/molecules/reply/reply-content';
 import { Reply } from '@/types';
 import { forwardRef, useState } from 'react';
+import { ReplyContent } from './reply-content';
 
 /** 대댓글 삭제를 위해 ReplyContent를 둘러싸기 위해 만든 컴포넌트 */
 export const RereplyItem = forwardRef<HTMLLIElement, Reply>((props, ref) => {
